+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.test.run.R
@@ -1,6 +1,6 @@
 concerto.test.run <-
 function(testId, params=list(), extraReturns=c()) {
-    test <- concerto.test.get(testId, cache=F, includeSubObjects=T)
+    test <- concerto.test.get(testId, includeSubObjects=T)
     if (is.null(test)) stop(paste("Test #", testId, " not found!", sep = ''))
     concerto.log(paste0("running test #", test$id, ": ", test$name, " ..."))