+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * improved import comparison hashing
 * sessionHash property exposed in testRunner javascript object
 * assessment node fixes/improvements
+* concerto objects caching enabled for test session
 
 ## 5.0.12 (2020-04-07)